+++ src/main/java/blusunrize/immersiveengineering/client/gui/MixerScreen.java
@@ -46,6 +46,7 @@ public class MixerScreen extends IEContainerScreen<MixerContainer>
 		super(container, inventoryPlayer, title, TEXTURE);
 		this.tile = container.tile;
 		this.imageHeight = 167;
+		this.inventoryLabelY = this.imageHeight-91;
 	}
 
 	@Nonnull