+++ src/main/java/blusunrize/immersiveengineering/client/gui/SorterScreen.java
@@ -47,6 +47,7 @@ public class SorterScreen extends IEContainerScreen<SorterContainer>
 		super(container, inventoryPlayer, title, TEXTURE);
 		this.tile = container.tile;
 		this.imageHeight = 244;
+		this.inventoryLabelY = this.imageHeight-91;
 	}
 
 	@Override